Remove development database URL from .env
The development database URL was set in .env and that meant the
development database was the default for all environments, including
test.  Although convenient for new contributors, it is confusing when
development database results are returned when tests run.  Also, it is
possible that local data might cause tests to fail or pass incorrectly.